Started debugging.

+++ thermo_functions.cpp
@@ -4,7 +4,7 @@
 
 const double CELS_TO_KELVIN = 273.15;					// Conversion factor from [C] to [K]
 const double KPa_TO_BAR = 100;							//conversion from [kPa] to [bar]
-const double CRIT_TEMP_NOS = 309.55;						// Critical Temperature of NOS [K]
+const double CRIT_TEMP_NOS = 309.55;					// Critical Temperature of NOS [K]
 const double CRIT_PRES_NOS = 7255;						// Critical Pressure of NOS [kPa]
 const double R_CONSTANT = 0.008314;						// Universal Gas Constant [kJ/(mol*K)]
 const double MM_NOS = 0.044013;							// Molar Mass of Nitrous Oxide [kg/mol]
@@ -95,8 +95,8 @@ double nox_vp(double T_Kelvin)
 	const float p[4] = { 1.0f, 1.5f, 2.5f, 5.0f };
 	const float b[4] = { -6.71893f, 1.35966f, -1.3779f, -4.051f };
 	double Tr = reduced_temperature(T_Kelvin);
-	float rab = 1.0 - Tr;
-	float shona = 0.0;
+	double rab = 1.0 - Tr;
+	double shona = 0.0;
 	for (int dd = 0; dd < 4; dd++)
 		shona += b[dd] * pow(rab, p[dd]);
 	double  Pvap= (CRIT_PRES_NOS/KPa_TO_BAR) * exp((shona / Tr));
@@ -109,7 +109,7 @@ double nox_enthV(double T_Kelvin)
 	const float bV[5] = { -200.0f, 440.055f, -459.701f, 434.081f, -485.338f };
 	double shonaL, shonaV;
 	double Tr = reduced_temperature(T_Kelvin);
-		float rab;
+		double rab;
 	rab = 1.0 - Tr;
 	shonaL = bL[0];
 	shonaV = bV[0];
